sami-generator/requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
google-cloud-texttospeech>=2.16.0
google-auth>=2.28.0
python-dotenv>=1.0.0
